Switching to work on sleep modes instead

+++ Src/peripherals/IMU/ASM330LHH.h
@@ -177,7 +177,7 @@ float imu_readGyro_Z();
 float imu_readTemp();
 
 //If there is no comments here, yell at me in PR
-void imu_continuous_dma(uint16_t rx_buffer[7]);
+void imu_continuous_dma(uint8_t rx_buffer[7]);
 
 /**
  * Tests SPI communication using WHO_AM_I register
